refactor(cli): split run.go into run, run_finish, and run_targets
